+++ gems/aws-sdk-s3/CHANGELOG.md
@@ -1,6 +1,8 @@
 Unreleased Changes
 ------------------
 
+* Issue - Ensure object metadata is not lost on multipart copy (#2821).
+
 1.119.0 (2023-01-26)
 ------------------